requirements.txt
anyio==3.6.2
autopep8==2.0.1
click==8.1.3
colorama==0.4.6
fastapi==0.89.1
FastAPI-SQLAlchemy==0.2.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pydantic==1.10.4
python-dotenv==0.21.1
sniffio==1.3.0
SQLAlchemy==1.4.19
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
Werkzeug==2.2.2
fastapi_pagination==0.11.3
PyJWT==2.0.0
python-multipart==0.0.5
passlib[bcrypt]